requirements.txt
# Enhanced Real-Time Billing Audit System Requirements
# Core dependencies for Excel generation and audit functionality

# Essential Smartsheet Integration
smartsheet-python-sdk>=2.105.1

# Excel Generation and Data Processing
openpyxl>=3.0.9
pandas>=1.3.0
python-dateutil>=2.8.2

# Environment Configuration
python-dotenv>=0.19.0

# HTTP Requests and API Handling
requests>=2.25.1

# Image Processing (for logos)
Pillow>=8.3.0

# Optional: Enhanced Analytics (will be skipped in ultra-light mode)
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.0.0

# Optional: System Information and ML (for advanced features)
psutil>=5.8.0
scikit-learn>=1.0.0
joblib>=1.1.0
networkx>=2.6.0

# Optional: Advanced AI capabilities (CPU-optimized for GitHub Actions)
torch --index-url https://download.pytorch.org/whl/cpu
tensorflow-cpu>=2.8.0
transformers[torch]>=4.15.0
spacy>=3.4.0
